🐛 [fix]: 자소서 경험카드 깜빡임 오류 해결, 내 포인트 UI 수정
