+++ src/app/components/mypage/MyPointContainer.tsx
@@ -3,6 +3,8 @@
 import { useEffect, useState } from 'react'
 import Link from 'next/link'
 import PointCard from './PointCard'
+import Icons from '../common/Icons'
+import { pointW } from '@/app/ui/IconsPath'
 
 interface UserInfoProps {
   email: string
@@ -84,17 +86,17 @@ const MyPointContainer = () => {
           <div className="w-[100px] h-[0px] left-0 top-[-0.5px] absolute border-2 border-gray-800" />
         </div>
         <div className="h-[39px] relative">
-          <div className="w-[1000px] h-[27px] left-[50px] top-0 absolute gap-[218px] items-center inline-flex">
-            <div className="text-black text-lg font-semibold leading-[27px]">
+          <div className="w-full h-[27px]  top-0 absolute gap-[218px] items-center inline-flex">
+            <div className="text-black text-lg font-semibold leading-[27px] absolute left-[8px] w-[170px] text-center">
               사용 일시
             </div>
-            <div className="text-black text-lg font-normal leading-[27px]">
+            <div className="text-black text-lg font-normal leading-[27px] absolute left-[300px] w-[150px] text-center">
               사용 포인트
             </div>
-            <div className="text-black text-lg font-normal leading-[27px]">
+            <div className="text-black text-lg font-normal leading-[27px]  text-center absolute left-[618px] w-[100px]">
               사용 유형
             </div>
-            <div className="text-black text-lg font-normal leading-[27px]">
+            <div className="text-black text-lg font-normal leading-[27px] text-center absolute left-[914px] w-[100px]">
               보유 포인트
             </div>
           </div>
@@ -114,9 +116,9 @@ const MyPointContainer = () => {
       <div className="left-[201px] top-[295px] absolute text-black text-xl font-bold leading-[30px]">
         <Link href="/mypage/pointcharge">충전 내역</Link>
       </div>
-      <div className="w-52 h-[54.45px] left-[960px] top-[165px] absolute">
-        <div className="left-[71.16px] top-[10.93px] absolute text-white text-xl font-semibold leading-[30px]">
-          충전하기
+      <div className="w-52 h-[54.45px] left-[900px] top-[160px] absolute items-center justify-center">
+        <div className="w-[200px] left-0 top-[10.93px] absolute h-[50px] items-center justify-center text-white border-2 border-white text-xl font-semibold leading-[30px] rounded-[30px] inline-flex">
+        <Icons className="mt-2" name={pointW} />충전하기
         </div>
       </div>
       <div className="w-[105.75px] h-[18px] left-[75px] top-[82.68px] absolute text-gray-900 text-[28px] font-bold font-['Rubik'] leading-[30px]">
