+++ src/app/ui/svg/ai/AI1.tsx
@@ -1,3 +1,5 @@
+import React from 'react';
+
 const AI1 = () => {
   return (
     <svg
