+++ src/app/ui/svg/ai/AI2.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 const AI2 = () => {
   return (
     <svg
