+++ src/app/ui/svg/application/Application3.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 const Application3 = () => {
   return (
     <svg
