+++ src/app/ui/svg/back/Back1.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 const Back1 = () => {
   return (
     <svg
