+++ src/app/ui/svg/design/Design1.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 const Design1 = () => {
   return (
     <svg
