+++ src/app/ui/svg/design/Design3.tsx
@@ -1,3 +1,5 @@
+
+import React from 'react';
 const Design3 = () => {
   return (
     <svg
