+++ src/app/ui/svg/web/Web1.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 const Web1 = () => {
   return (
     <svg
