+++ src/app/ui/svg/web/Web2.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 const Web2 = () => {
   return (
     <svg
