+++ src/app/ui/svg/web/Web3.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 const Web3 = () => {
   return (
     <svg
